Fix the page for the bibliography ToC entry.
If there is additional material between \backmatter and \bibliography,
then the page number for the bibliography is likely to be wrong.
\backmatter inserts a \clearpage which forces all floats to flush in
addition to ending the page. Thus,
    \backmatter
    \bibliography{references}
works correctly, but omitting \backmatter or adding additional
material will frequently give the wrong page number. Adding \clearpage
to the beginning of \thebibliography fixes the issue.